Refactor to use mix and match parse functions with site configs in JSON instead of a class for every site.
