+++ data/target_sites.json
@@ -1,0 +1,57 @@
+{
+  "nyt": {
+    "site": "nyt",
+    "suffix": "?pagewanted=all",
+    "domains": ["www.nytimes.com"],
+    "feeder_pattern": "^https?://www.nytimes.com/202",
+    "feeder_pages": [
+        "https://www.nytimes.com/",
+        "https://www.nytimes.com/pages/world/",
+        "https://www.nytimes.com/pages/national/",
+        "https://www.nytimes.com/pages/todayspaper/",
+        "https://www.nytimes.com/pages/politics/",
+        "https://www.nytimes.com/pages/nyregion/",
+        "https://www.nytimes.com/pages/business/",
+        "https://www.nytimes.com/pages/technology/",
+        "https://www.nytimes.com/pages/sports/",
+        "https://dealbook.nytimes.com/",
+        "https://www.nytimes.com/pages/science/",
+        "https://www.nytimes.com/pages/health/",
+        "https://www.nytimes.com/pages/arts/",
+        "https://www.nytimes.com/pages/style/",
+        "https://www.nytimes.com/pages/opinion/",
+        "https://www.nytimes.com/pages/automobiles/",
+        "https://www.nytimes.com/pages/books/",
+        "https://www.nytimes.com/crosswords/",
+        "https://www.nytimes.com/pages/dining/",
+        "https://www.nytimes.com/pages/education/",
+        "https://www.nytimes.com/pages/fashion/",
+        "https://www.nytimes.com/pages/garden/",
+        "https://www.nytimes.com/pages/magazine/",
+        "https://www.nytimes.com/pages/business/media/",
+        "https://www.nytimes.com/pages/movies/",
+        "https://www.nytimes.com/pages/arts/music/",
+        "https://www.nytimes.com/pages/obituaries/",
+        "https://www.nytimes.com/pages/t-magazine/",
+        "https://www.nytimes.com/pages/arts/television/",
+        "https://www.nytimes.com/pages/theater/",
+        "https://www.nytimes.com/pages/travel/",
+        "https://topics.nytimes.com/top/opinion/thepubliceditor/"
+    ],
+    "use_archive": true,
+    "parser_name": "nyt_style"
+  },
+  "eurogamer": {
+    "site": "eurogamer",
+    "domains": ["www.eurogamer.net"],
+    "feeder_pattern": "^https?://www.eurogamer.net/",
+    "feeder_pages": [
+        "https://www.eurogamer.net/",
+        "https://www.eurogamer.net/latest",
+        "https://www.eurogamer.net/pc",
+        "https://www.eurogamer.net/nintendo"
+    ],
+    "use_archive": false,
+    "parser_name": "article_based"
+  }
+}
